requirements.txt
pyrogram==1.4.16
tgcrypto
yt-dlp
python-dotenv
requests
asgiref
deezer-python
spotipy

Pillow
ffmpeg-python
youtube-dl
wget

youtube-search-python==1.4.9
youtube-search
dnspython
gitpython
aiofiles
aiohttp
requests
motor
psutil
future
